# AnsibleRepositoryController: speed up tests.