Speed up AnsibleRepositoryController tests:
  * do not recreate `:embedded_ansible_configuration_script_source` for each test.
  * do not run test 2x just to test the http_code

Generally I don't see why one should re-create DB entities and re-render a page to test trivial individual properties of the page one by one.

By the way **testing just the HTTP code makes a test pointless** because w/o `bypass_rescue` you get a `200 OK` even when an exception is raised.

Before:
```
Finished in 4.3 seconds (files took 9.26 seconds to load)
18 examples, 0 failures
```

After:
```
Finished in 2.33 seconds (files took 9.2 seconds to load)
14 examples, 0 failures
```