@martinpovolny 2018-11-23
ping @mzazrivec, @himdel